Implemented TestColObserver, TestInputObserver, GridManager and partially implemented Automated Bomb Dropping
Added a few test observer classes to test prototypes of reactions in a controlled fashioned. For example, I can now use a key stroke to force the grid to drop a bomb or a ship to respawn after being destroyed. Also partially automated the grid to drop a bomb based on time via the grid manager - I would like to refactor this in order to make the bomb drop rate dependent on the number of steps taken by the grid.